[#158] Refactored Graph Components
* Limite number of decimals in GraphComponent to 2
* Moved all graph related components to GraphComponents/
* Updated path in DisplaySensor and Comparator
