+++ frontend/src/components/GraphComponents/GraphControl.css
